requirements.txt
Django==1.8.3
gunicorn==19.3.0
numpy==1.9.2
python-dateutil==2.4.2
requests==2.7.0
six==1.9.0
virtualenv==13.1.0
virtualenvwrapper-powershell==12.7.8
wheel==0.24.0
